build(docker): add build stage to Dockerfile
Adds a build stage to the Dockerfile to make it easier to onboard non-Java developers. The build is self-contained and only requires the src directory to be COPYed into the image. The build stage is used to compile the Java code and run the tests. The final image is built from the build stage and only contains the compiled code and the necessary dependencies.